# REF: use public pandas API in dataframe.empty