The current usage of non-public API is breaking the docbuild on a pandas PR: https://github.com/pandas-dev/pandas/pull/40149/checks?check_run_id=2102847175#step:5:152

See also: https://github.com/pandas-dev/pandas/issues/40226

A slightly more performant implementation may become possible following https://github.com/pandas-dev/pandas/issues/39776