@jbrockmendel 2021-03-15
Yep, thats a pretty big slowdown.  I'll see what I can do upstream, and take another try at this in the interim.