@martindurant 2021-03-15
Thanks, @jbrockmendel !

Interestingly, I [wrote about this](http://martindurant.github.io/blog/pandas-and-memory-allocation/) all the way back in 2017. @jreback helped write this initially, and @TomAugspurger has helped keep this module in sync with pandas changes.