@jbrockmendel 2021-03-15
Updated with an implementation that slightly outperforms the status quo

```
In [3]: %timeit out = fastparquet.dataframe.empty('i4,u2,f4,f2,f4,M8,category', 1000000, cats={6: 5})
741 µs ± 11.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- main
712 µs ± 4.04 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <-- PR
```

This still accesses internals in a way that is not officially supported, but should be at least future-proof to https://github.com/pandas-dev/pandas/pull/40149 (though not ArrayManager)