@jbrockmendel 2021-03-28
Some of what I'm working on in pandas would be simplified if this made it into a released version of fastparquet.  Is that on the horizon?